Move LocalAnalyzer to rustc_mir::ssa_analyze
